Uso do novo endpoint para guias para insucesso
